Add TailwindCSS as CSS framework
We use a framework to increase the speed of development. Tailwind was
selected because it's utility based philosophy gets out of the way of
getting things done.

We begin by cleaning up the logo to be a reusable component and avoiding
all that SVG in the index page file.
