+++ .browserslistrc
@@ -1,0 +1,1 @@
+> 1%, last 2 versions, not dead
