+++ gatsby-config.js
@@ -69,6 +69,7 @@ module.exports = {
         ],
       },
     },
+    `gatsby-plugin-postcss`,
     {
       resolve: `gatsby-plugin-manifest`,
       options: {
