+++ package.json
@@ -23,8 +23,12 @@
     "react-helmet": "^5.2.1"
   },
   "devDependencies": {
+    "autoprefixer": "^9.7.4",
+    "gatsby-plugin-postcss": "^2.1.19",
     "gh-pages": "^2.1.1",
-    "prettier": "^1.18.2"
+    "postcss-import": "^12.0.1",
+    "prettier": "^1.18.2",
+    "tailwindcss": "^1.1.4"
   },
   "scripts": {
     "build": "gatsby build",
