+++ postcss.config.js
@@ -1,0 +1,11 @@
+const postcssImport = require('postcss-import');
+const tailwind = require('tailwindcss');
+const autoprefixer = require('autoprefixer');
+
+module.exports = {
+  plugins: [
+    postcssImport,
+    tailwind,
+    autoprefixer,
+  ],
+};
