+++ src/components/Layout.js
@@ -9,7 +9,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 // import { useStaticQuery, graphql } from 'gatsby';
 
-import './layout.css';
+import '../styles/index.css';
 
 const Layout = ({ children, isFront = false }) => {
   // const data = useStaticQuery(graphql`
