+++ src/components/atoms/Logo.js
@@ -1,0 +1,223 @@
+import React from "react";
+
+export default function Logo({ width = "200", height = "200" }) {
+  return <svg
+    xmlns="http://www.w3.org/2000/svg"
+    version="1"
+    id="_x30_"
+    viewBox="0 0 53 53"
+    width={width}
+    height={height}
+  >
+    <style id="style2">{`
+          /* Red */
+          .st2 {
+              fill: rgb(255, 60, 60);
+          }
+
+          .st3 {
+              fill: rgb(255, 0, 0);
+          }
+
+          /* White */
+          .st0 {
+              fill: rgb(255, 255, 255);
+          }
+          .st1 {
+              fill: rgb(240, 240, 240);
+          }
+
+          /* Blue */
+          .st6 {
+              fill: rgb(60, 60, 255)
+          }
+
+          .st7 {
+              fill: rgb(0, 0, 255);
+          }
+
+          /* Orange */
+          .st4 {
+              fill: rgb(255,99,20);
+          }
+
+          .st5 {
+              fill: rgb(255,79,0);
+          }
+        `}</style>
+    <g id="SPLINE_11_" transform="translate(0 -8) scale(.08653)">
+      <g id="XMLID_1_">
+        <g id="g60">
+          <path
+            className="st0"
+            d="M162 596l-33 33-6-1-6-1-5-2-3-1-2-1-3-1-2-1-2-1-3-2-2-1-2-1-2-2-1-2-2-2-2-2-1-1-2-3-1-2-1-2-1-3-1-2-1-3-1-2-2-6-1-6-1-6 33-33c15 22 34 41 56 57z"
+            id="path4"
+            fill="#44f0d1"
+          ></path>
+          <path
+            className="st0"
+            d="M157 489l-51 50c-22-30-37-66-43-103h72c4 19 11 36 22 53z"
+            id="path6"
+            fill="#44f0d1"
+          ></path>
+          <path
+            className="st0"
+            d="M549 357h-72c-4-19-12-37-22-53l51-51c22 30 37 66 43 104z"
+            id="path8"
+            fill="#44f0d1"
+          ></path>
+          <path
+            className="st0"
+            d="M538 214l1 6-33 33c-16-22-35-40-56-57l33-33 6 1 6 1 5 2 3 1 2 1 3 1 2 1 2 1 3 2 2 1 2 2 2 1 1 2 2 2 2 1 1 3 2 2 1 2 1 2 1 2 1 3 1 3 1 2 2 6z"
+            id="path10"
+            fill="#44f0d1"
+          ></path>
+          <path
+            className="st1"
+            d="M213 545l-51 51c-22-16-41-35-56-57l51-50c14 22 33 42 56 56z"
+            id="path12"
+            fill="#09d4b0"
+          ></path>
+          <path
+            className="st1"
+            d="M506 253l-50 50c-14-22-34-42-57-56l51-51c22 16 41 35 56 57z"
+            id="path14"
+            fill="#09d4b0"
+          ></path>
+          <path
+            className="st2"
+            d="M346 639v47l-5 4-5 3-5 3-5 2-5 2-5 1-5 1h-10l-5-1-5-2-5-1-5-3-5-2-5-4-5-3v-47c26 4 53 4 80 0z"
+            id="path16"
+            fill="#7f7fe4"
+          ></path>
+          <path
+            className="st2"
+            d="M266 567v72c-37-6-73-21-104-43l51-51c16 10 34 18 53 22z"
+            id="path18"
+            fill="#7f7fe4"
+          ></path>
+          <path
+            className="st2"
+            d="M450 196l-51 51c-16-10-34-18-53-22v-72c37 6 73 21 104 43z"
+            id="path20"
+            fill="#7f7fe4"
+          ></path>
+          <path
+            className="st3"
+            d="M346 225l-2 5-3 5-2 4-3 4-2 4-3 3-2 4-3 2-2 3-3 2-2 2-3 2-2 1-3 1-2 1h-3l-2-1h-3l-2-1-3-1-2-2-3-2-2-2-3-3-2-3-3-3-2-3-3-4-2-4-3-5-2-4-2-5c28-8 54-8 79 0zm-28 13c1-7-5-13-12-13s-12 6-12 13a13 13 0 0 0 25 0z"
+            id="path22"
+            fill="#5252c9"
+          ></path>
+          <path
+            className="st3"
+            d="M343 562l3 5c-27 7-56 8-80 0l2-5 3-5 2-4 3-4 2-4 3-3 2-4 3-2 2-3 3-2 2-2 3-2 2-1 3-1h2l3-1h2l3 1 2 1 3 1 2 2 3 2 2 2 3 3 2 2 3 4 2 3 3 4 2 4 3 4zm-24-8a13 13 0 1 0-26 0 13 13 0 0 0 26 0z"
+            id="path24"
+            fill="#5252c9"
+          ></path>
+          <path
+            className="st3"
+            d="M346 567v72c-27 4-54 4-80 0v-72c26 6 53 6 80 0z"
+            id="path26"
+            fill="#5252c9"
+          ></path>
+          <path
+            className="st3"
+            d="M346 225c-26-6-53-6-80 0v-72c27-4 54-4 80 0z"
+            id="path28"
+            fill="#5252c9"
+          ></path>
+          <path
+            className="st4"
+            d="M539 573l-1 6-1 6-2 5-1 3-1 2-1 3-1 2-1 2-2 3-1 2-1 2-2 2-2 1-2 2-2 2-1 1-3 2-2 1-2 1-3 1-2 1-3 1-2 1-6 2-6 1-6 1-33-33c22-15 41-34 57-56z"
+            id="path30"
+            fill="#e6da6e"
+          ></path>
+          <path
+            className="st4"
+            d="M399 545l50 51c-30 22-66 37-103 43v-72c19-4 36-12 53-22z"
+            id="path32"
+            fill="#e6da6e"
+          ></path>
+          <path
+            className="st4"
+            d="M267 153v72c-19 4-37 11-54 21l-50-50c29-21 64-36 100-42l4-1z"
+            id="path34"
+            fill="#e6da6e"
+          ></path>
+          <path
+            className="st4"
+            d="M130 163l33 33c-21 15-41 34-57 56l-33-33 1-6 1-6 2-5 1-3 1-2 1-3 1-2 1-2 2-3 1-2 1-2 2-2 2-1 2-2 1-2 2-1 3-2 2-1 2-1 2-1 3-1 3-1 2-1 6-2 5-1z"
+            id="path36"
+            fill="#e6da6e"
+          ></path>
+          <path
+            className="st5"
+            d="M455 489l51 51c-16 22-35 41-57 56l-50-51c22-14 42-33 56-56z"
+            id="path38"
+            fill="#cabd4b"
+          ></path>
+          <path
+            className="st5"
+            d="M163 196l50 51c-22 14-42 33-56 56l-51-51c16-22 35-41 57-56z"
+            id="path40"
+            fill="#cabd4b"
+          ></path>
+          <path
+            className="st6"
+            d="M106 252l51 51c-10 16-18 34-22 53H63c6-37 21-73 43-104z"
+            id="path42"
+            fill="#66a2f6"
+          ></path>
+          <path
+            className="st6"
+            d="M16 356h47c-4 26-4 53 0 80H16l-4-5-3-5-3-5-2-5-2-5-1-5-1-5v-10l1-5 1-5 2-5 3-5 2-5 4-5z"
+            id="path44"
+            fill="#66a2f6"
+          ></path>
+          <path
+            className="st6"
+            d="M506 540l-51-51c10-16 18-34 22-53h72c-6 37-21 73-43 104z"
+            id="path46"
+            fill="#66a2f6"
+          ></path>
+          <path
+            className="st6"
+            d="M612 391v10l-1 5-2 5-1 5-3 5-2 5-4 5-3 5h-47c3-26 3-52 0-79h47l4 4 3 5 3 5 2 5 2 5 1 5z"
+            id="path48"
+            fill="#66a2f6"
+          ></path>
+          <path
+            className="st7"
+            d="M140 433l-5 3c-6-27-8-55 0-80l5 2 4 3 5 2 4 3 4 2 3 3 3 2 3 3 3 2 2 3 2 2 2 3 1 2 1 3 1 2v3l-1 2v3l-1 2-1 3-2 2-2 3-2 2-3 3-3 2-3 3-3 2-4 3-4 2-5 3zm20-37c0-7-5-12-12-12s-13 5-13 12 6 13 13 13 12-6 12-13z"
+            id="path50"
+            fill="#247eff"
+          ></path>
+          <path
+            className="st7"
+            d="M63 356h72c-5 26-4 52 0 80H63c-4-27-4-54 0-80z"
+            id="path52"
+            fill="#247eff"
+          ></path>
+          <path
+            className="st7"
+            d="M477 357c12 26 12 53 0 79l-5-2-4-3-5-2-4-3-4-2-3-3-3-2-3-3-3-2-2-3-2-2-2-3-1-2-1-3v-2l-1-3 1-2v-3l1-2 1-3 2-2 2-3 2-2 3-3 3-2 3-3 3-2 4-3 4-2 5-3 4-2zm0 39a13 13 0 1 0-25 0 13 13 0 0 0 25 0z"
+            id="path54"
+            fill="#247eff"
+          ></path>
+          <path
+            className="st7"
+            d="M549 357c4 26 4 53 0 79h-72c6-26 6-53 0-79z"
+            id="path56"
+            fill="#247eff"
+          ></path>
+          <path
+            className="st2"
+            d="M346 106v47c-27-4-52-4-80 0v-47l2-2 2-1 1-2 3-1 2-2 2-1 2-1 3-2h2l3-1 3-1 3-1h6l6-1 6 1h7l2 1 3 1 3 1 3 1 2 1 2 1 3 1 2 2 2 1 2 2 2 2z"
+            id="path58"
+            fill="#7f7fe4"
+          ></path>
+        </g>
+      </g>
+    </g>
+  </svg>;
+}
